Split misc test from test_math.py
